backend/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.26.0
python-multipart==0.0.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
openai==1.3.7
cloudinary==1.36.0
supabase==1.2.0
Pillow==10.1.0
pytest==7.4.3
black==23.11.0
ruff==0.1.5
pytest-cov==4.1.0 